Add AvbTest for Android Verified Boot app
This is a "white box" test. That is, it requires some control
over the internals of the AVB app as it's running on Nugget OS.

There are certain actions that are only allowed when the AP is in
the BIOS, and others that are only allowed when it is in the
high-level OS. In a running system, the AVB app watches the state
of the AP to determine when it has rebooted and thus entered
BIOS. It relies on the BIOS to inform the app before launching
the HLOS.

We don't (yet?) have the ability to force the AP into each state
as part of this test, or to fake the MSM_RST_OUT_L signal on a
standalone Citadel test board. Instead, this test requires that
the running Nugget OS image be compiled ("make TEST_IMAGE=1 ...")
to add a special command that will force AVB to believe that the
AP is in the BIOS.

Change-Id: I46d1cd17594929cc185f21e7e48553bcd6f83560