avb_tests: allow 0 or 1 for the selector
In order to support the selector increment the server will need, the
keys were labeled (proto version + key version). 2 will be the prod
key and 0 is the test key, but we were using 1 as a placeholder.
This change papers over it until we decide on a enum naming scheme
for ResetToken.

Test: touch avb_app.c; change the defines; run the tests
Bug: 77300593
Change-Id: I858e301631976dd316ff18d8ea05e7f783bf6047