avb_tests: split out reset keys; enable key testing
This change uses ProductionResetTest() to verify server signatures
are valid with AVB's internal keys and implementation.

The server signs a nonce=0,hash=00.00 which acts as
the "test input". This signature is put into the test-data
reset_key_data.cc file.

AVB uses a selector to identify key/protocol version:
  0 = test, 1 = server-dev, 2 = server-prod

Currently 1 is the default and 0 is what TEST_IMAGE produces.

This test supports testing whichever key is active on the image.
Many AVB tests expect selector==1, so they will fail if run with
a different key, but this test may be run explicitly using
   bazel run runtests -- \
     --gtest_filter=AvbTest.ResetKeyNullTokenSignatureTest

To switch the build, the RESET_KEY macro can be used on BOARD=red:
      touch user/avb/avb_app.c
      make BOARD=red ... RESET_KEY=1  # 0 or 2

For other builds, the key will default to what is in the define.

Test: Followed the instructions above for keys 0, 1, 2.
Bug: 74946926

Change-Id: I4a7b67f9bc2ab14731f4340d56382df22cd588b2