keymaster:  add cert recovery test
This test loads PVT certs onto a PVT
chip, and ensures that they are loaded
correctly.

NOTE: test does not request attestation,
as HAL code is unavailable in this environment.
However, this test still provides good confidence
that the provided cert packet was acceptable, as
it must have decrypted correctly for the test to pass.

TODO: port corresponding test over to jenkins.

Bug: 111997966
Test: ProvisionCertificates test passes at desk
Change-Id: I1c331c54e20297dc5d13256058c5d9e51c99d36d